Version: 1.1
Description: A general contact form with add fields and a shortcode to
be used on page or post. The fields type can be text,text area,radio
buttom,dropdown,password. The form can also be used as a widget. Also,
to enhance the contact form usability, a security code feature(captcha)
is also added and all the form details can be viewed at the admin side.